Fix crud bugs

+++ WeTapAPI/Application/Features/Users/Commands/DeleteUser/DeleteUserCommand.cs
@@ -2,4 +2,4 @@
 
 namespace Application.Features.Users.Commands.DeleteUser;
 
-public record DeleteUserCommand(int Id) : IRequest;
+public record DeleteUserCommand(long Id) : IRequest;
